refactor(runtimed): share snapshot widget comm projection
